Add pavement script in respective folders, deleted old pavement files, updated readme to reflect support for Python3.6+
