+++ ios/examples/run-parallel-tests/features/environment.py
@@ -16,7 +16,7 @@ def before_feature(context, feature):
     desired_capabilities = CONFIG['capabilities']
     desired_capabilities['device'] = CONFIG['environments'][TASK_ID]['device']
     context.browser = webdriver.Remote(
-        desired_capabilities=desired_capabilities,
+        desired_capabilities=dict(desired_capabilities),
         command_executor="http://%s:%s@hub-cloud.browserstack.com/wd/hub" % (BROWSERSTACK_USERNAME, BROWSERSTACK_ACCESS_KEY)
     )
